requirements.txt
backoff==1.8.0
certifi==2022.6.15
charset-normalizer==2.1.0
ciso8601==2.2.0
idna==3.3
jsonschema==2.6.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
simplejson==3.11.1
singer-python==5.12.2
six==1.16.0
urllib3==1.26.10
